Start refactoring warning diagnostics to be program-independent
